requirements.txt
flask
transformers
opencv-python
Pillow
torch
